+++ reactions/sdk/javascript/.npmignore
@@ -1,0 +1,5 @@
+examples/
+test/
+node_modules/
+.DS_Store
+.env
