+++ reactions/sdk/javascript/examples/advanced-ts/Dockerfile
@@ -1,15 +1,10 @@
-FROM node:20
+FROM node:20-alpine
 
 WORKDIR /usr/src/app
 
 COPY . .
 
 RUN npm install -g typescript
-RUN npm install
-RUN npm run build
-
-WORKDIR /usr/src/app/examples/simple
-
 RUN npm install
 RUN npm run build
