+++ reactions/sdk/javascript/examples/advanced-ts/reaction.yaml
@@ -3,6 +3,8 @@ apiVersion: v1
 name: test-advanced
 spec:
   kind: Advanced
+  properties:
+    MyConnectionString: "some connection string"
   queries:
     query1: |
       greeting: "Hello, World!"
