+++ reactions/sdk/javascript/examples/simple-ts/.dockerignore
