+++ reactions/sdk/javascript/examples/simple-ts/.gitignore
