+++ reactions/sdk/javascript/examples/simple-ts/Dockerfile
@@ -1,15 +1,10 @@
-FROM node:20
+FROM node:20-alpine
 
 WORKDIR /usr/src/app
 
 COPY . .
 
 RUN npm install -g typescript
-RUN npm install
-RUN npm run build
-
-WORKDIR /usr/src/app/examples/advanced
-
 RUN npm install
 RUN npm run build
