+++ reactions/sdk/javascript/examples/simple-ts/reaction.yaml
