+++ reactions/sdk/javascript/examples/simple-ts/tsconfig.json
