--- reactions/sdk/javascript/examples/simple/src/index.ts
@@ -1,25 +1,0 @@
-import { DrasiReaction, ChangeEvent } from '@drasi/reaction-sdk';
-
-async function main() {
-
-    let myReaction = new DrasiReaction(async (event: ChangeEvent) => {
-        
-        console.log(`Received change sequence: ${event.sequence} for query ${event.queryId}`);
-        
-        for (let added of event.addedResults) {
-            console.log(`Added result: ${JSON.stringify(added)}`);
-        }
-        
-        for (let deleted of event.deletedResults) {
-            console.log(`Removed result: ${JSON.stringify(deleted)}`);
-        }
-        
-        for (let updated of event.updatedResults) {
-            console.log(`Updated result - before: ${JSON.stringify(updated.before)}, after: ${JSON.stringify(updated.after)}`);
-        }
-    });
-
-    await myReaction.start();
-}
-
-main().catch(console.error);
