+++ reactions/sdk/javascript/src/types/ControlEvent.d.ts
@@ -5,6 +5,9 @@
  * and run json-schema-to-typescript to regenerate this file.
  */
 
+/**
+ * An event that represents a control signal from the query, such are start, stop, etc.
+ */
 export type ControlEvent = ResultEvent & {
   kind: "control";
   controlSignal: ControlSignal;
@@ -12,8 +15,17 @@ export type ControlEvent = ResultEvent & {
 };
 
 export interface ResultEvent {
+  /**
+   * The ID of the query that the event originated from
+   */
   queryId: string;
+  /**
+   * The sequence number of the event
+   */
   sequence: string;
+  /**
+   * The time at which the source change was recorded
+   */
   sourceTimeMs: string;
   metadata?: RecordUnknown;
   [k: string]: unknown;
